refactor(streamlit): extract Settings tab and data tab runners from app.py (#118)
- Add app_settings_ui.render_settings_tab for persistence, map/table forms, taxonomy, data path
- Add _run_non_map_data_tab_fragments for Checklist/Rankings/Yearly/Country/Maintenance
- Trim app.py imports moved to the settings module

Behaviour unchanged; further extractions (prep spinner, landing) can follow in smaller PRs.

Refs #118

Made-with: Cursor
